requirements/project.txt
Mezzanine==0.12.4
pillow
django_stackato
